requirements.txt
sqlalchemy
google-cloud-language